# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Satcop Smart Visitor API",
    "description": "MVP FastAPI stub + Priority P2 Pickup (P1\u2013P6+H1) + after-hours (A1\u2013A6) + escort/zones Access Rules (B4) + Emergency Blast (E3 / B1\u2013B6). Contract \u00a7\u00a70\u20135 visit lifecycle unchanged. SMS mock / WA skipped_hold. Demo watermark; no production / live-school deploy.",
    "version": "0.6.0"
  },
  "paths": {
...
    "/v1/schools/": {
      "get": {
        "tags": [
          "schools"
        ],
        "summary": "List Schools Admin",
        "description": "List all schools \u2014 admin of any school (ops bootstrap helper).",
        "operationId": "list_schools_admin_v1_schools__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "schools"
        ],
        "summary": "Create School",
        "description": "Create a school tenant + admin bootstrap.\n\nPrefer admin of any school (role check only). Mints SCH-<SLUG>-01,\ndefault 4 gates, Mon\u2013Fri 08\u201318 hours, escort/zone defaults.\nNever overwrites SCH-DEMO-01.",
        "operationId": "create_school_v1_schools__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/schools": {
      "get": {
        "tags": [
          "schools"
        ],
        "summary": "List Schools Admin",
        "description": "List all schools \u2014 admin of any school (ops bootstrap helper).",
        "operationId": "list_schools_admin_v1_schools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "schools"
        ],
        "summary": "Create School",
        "description": "Create a school tenant + admin bootstrap.\n\nPrefer admin of any school (role check only). Mints SCH-<SLUG>-01,\ndefault 4 gates, Mon\u2013Fri 08\u201318 hours, escort/zone defaults.\nNever overwrites SCH-DEMO-01.",
        "operationId": "create_school_v1_schools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/schools/me/blast-config": {
      "get": {
        "tags": [
          "schools"
        ],
        "summary": "Get Blast Config",
        "operationId": "get_blast_config_v1_schools_me_blast_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "schools"
        ],
        "summary": "Put Blast Config",
        "operationId": "put_blast_config_v1_schools_me_blast_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlastConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "schools"
        ],
        "summary": "Patch Blast Config",
        "operationId": "patch_blast_config_v1_schools_me_blast_config_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlastConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlastConfigPatch": {
        "properties": {
          "emergencyBlastEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergencyblastenabled"
          },
          "blastStaffLaneEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blaststafflaneenabled"
          },
          "blastChannelsVisitor": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blastchannelsvisitor"
          },
          "blastChannelsStaff": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blastchannelsstaff"
          }
        },
        "type": "object",
        "title": "BlastConfigPatch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SchoolCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "schoolCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schoolcode"
          },
          "schoolId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schoolid"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Kolkata"
          },
          "adminUsername": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adminusername"
          },
          "adminDisplayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admindisplayname"
          },
          "adminPassword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adminpassword"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SchoolCreate",
        "description": "POST /v1/schools \u2014 mint tenant + admin bootstrap."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}